backend/requirements.txt
alembic==1.13.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
billiard==4.2.4
celery==5.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.4
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.20
google-auth==2.48.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
idna==3.11
Jinja2==3.1.6
kombu==5.6.2
kubernetes==29.0.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
oauthlib==3.3.1
orjson==3.11.7
packaging==26.0
passlib==1.7.4
prompt_toolkit==3.0.52
psycopg==3.2.13
psycopg-binary==3.2.13
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.3.0
python-multipart==0.0.22
PyYAML==6.0.3
redis==5.0.4
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.2
rich-toolkit==0.19.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
structlog==24.1.0
typer==0.21.1
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
uvicorn==0.30.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websocket-client==1.9.0
websockets==16.0
